# backend/docs/swagger.yaml
definitions:
  github_com_movie-matcher_backend_internal_models.AgeRating:
    enum:
    - g
    - pg
    - pg13
    - r
    - nc17
    - nc21
    - tvY
    - tvY7
    - tvG
    - tvPg
    - tv14
    - tvMa
    type: string
    x-enum-varnames:
    - AgeRatingG
    - AgeRatingPG
    - AgeRatingPG13
    - AgeRatingR
    - AgeRatingNC17
    - AgeRatingNC21
    - AgeRatingTVY
    - AgeRatingTVY7
    - AgeRatingTVG
    - AgeRatingTVPG
    - AgeRatingTV14
    - AgeRatingTVMA
...
  github_com_movie-matcher_backend_internal_models.Cast:
    properties:
      character:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Character'
      characterId:
        type: integer
      createdAt:
        type: string
      dubbings:
        description: 'Дубляж: даббер и язык'
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Dubbing'
        type: array
      episodeNumber:
        type: integer
      id:
        type: integer
      images:
        items: {}
        type: array
      person:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Person'
      personId:
        type: integer
      poster:
        description: 'CastPoster: изображение для карточки каста'
        type: string
      role:
        type: string
      roleType:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.RoleType'
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.Character:
    properties:
      avatar:
        type: string
      cast:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Cast'
        type: array
      createdAt:
        type: string
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      id:
        type: integer
      images:
        items: {}
        type: array
      name:
        type: string
      nameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.Dubbing:
    properties:
      cast:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Cast'
      castId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      language:
        description: 'Код языка: ru, en, ja и т.д.'
        type: string
      person:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Person'
      personId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.Genre:
    properties:
      createdAt:
        type: string
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      emoji:
        type: string
      id:
        type: integer
      mediaTypes:
        items: {}
        type: array
      name:
        type: string
      nameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      updatedAt:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_models.JSONMap:
    additionalProperties: true
    type: object
...
  github_com_movie-matcher_backend_internal_models.ListStatus:
    enum:
    - watching
    - planned
    - completed
    - onHold
    - dropped
    - rewatching
    type: string
    x-enum-varnames:
    - ListStatusWatching
    - ListStatusPlanned
    - ListStatusCompleted
    - ListStatusOnHold
    - ListStatusDropped
    - ListStatusRewatching
  github_com_movie-matcher_backend_internal_models.LocalizedString:
    additionalProperties:
      type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.MediaStaff:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      mediaId:
        type: integer
      mediaType:
        type: string
      person:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Person'
      personId:
        type: integer
      profession:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Profession'
      updatedAt:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_models.MediaStatus:
    enum:
    - announced
    - in_production
    - released
    - finished
    - cancelled
    - postponed
    type: string
    x-enum-comments:
      MediaStatusAnnounced: Анонсировано
      MediaStatusCancelled: Отменено
      MediaStatusFinished: Завершено
      MediaStatusInProduction: В производстве
      MediaStatusPostponed: Отложено
      MediaStatusReleased: Вышло / выходит
    x-enum-varnames:
    - MediaStatusAnnounced
    - MediaStatusInProduction
    - MediaStatusReleased
    - MediaStatusFinished
    - MediaStatusCancelled
    - MediaStatusPostponed
...
  github_com_movie-matcher_backend_internal_models.Person:
    properties:
      avatar:
        type: string
      biography:
        type: string
      biographyI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      birthDate:
        type: string
      cast:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Cast'
        type: array
      country:
        type: string
      createdAt:
        type: string
      firstName:
        type: string
      firstNameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      id:
        type: integer
      images:
        items: {}
        type: array
      lastName:
        type: string
      lastNameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      profession:
        items: {}
        type: array
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_models.ReviewStatus:
    enum:
    - neutral
    - positive
    - negative
    - surprised
    - disappointed
    - excited
    - confused
    - laugh
    - boring
    - cry
    - cool
    - cold
    - scary
    - shushing
    - exploding_head
    - face_holding_back_tears
    - kiss
    - pleading_face
    - saluting_face
    type: string
    x-enum-varnames:
    - ReviewStatusNeutral
    - ReviewStatusPositive
    - ReviewStatusNegative
    - ReviewStatusSurprised
    - ReviewStatusDisappointed
    - ReviewStatusExcited
    - ReviewStatusConfused
    - ReviewStatusLaugh
    - ReviewStatusBoring
    - ReviewStatusCry
    - ReviewStatusCool
    - ReviewStatusCold
    - ReviewStatusScary
    - ReviewStatusShushing
    - ReviewStatusExplodingHead
    - ReviewStatusFaceHoldingBackTears
    - ReviewStatusKiss
    - ReviewStatusPleadingFace
    - ReviewStatusSalutingFace
  github_com_movie-matcher_backend_internal_models.RoleType:
    enum:
    - main
    - supporting
    - cameo
    - extra
    - narrator
    - guest
    - background
    type: string
    x-enum-varnames:
    - RoleTypeMain
    - RoleTypeSupporting
    - RoleTypeCameo
    - RoleTypeExtra
    - RoleTypeNarrator
    - RoleTypeGuest
    - RoleTypeBackground
...
  github_com_movie-matcher_backend_internal_models.Studio:
    properties:
      country:
        type: string
      createdAt:
        type: string
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      id:
        type: integer
      name:
        type: string
      nameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      poster:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_models.TVSeries:
    properties:
      ageRating:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.AgeRating'
      cast:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Cast'
        type: array
      comments:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesComment'
        type: array
      createdAt:
        type: string
      currentEpisode:
        description: Текущая вышедшая серия
        type: integer
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      episodeDuration:
        type: integer
      episodesCount:
        type: integer
      favorites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesFavorite'
        type: array
      genres:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Genre'
        type: array
      id:
        type: integer
      images:
        items: {}
        type: array
      isHidden:
        type: boolean
      listStatus:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.ListStatus'
      lists:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesList'
        type: array
      poster:
        type: string
      rating:
        type: number
      ratingCount:
        type: integer
      releaseDate:
        type: string
      releaseSchedule:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.JSONMap'
      reviews:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesReview'
        type: array
      seasonNumber:
        type: integer
      similar:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeries'
        type: array
      sites:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesSite'
        type: array
      staff:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.MediaStaff'
        type: array
      status:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.MediaStatus'
      studios:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Studio'
        type: array
      themes:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Theme'
        type: array
      title:
        type: string
      titleI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      updatedAt:
        type: string
      videos:
        items: {}
        type: array
    type: object
  github_com_movie-matcher_backend_internal_models.TVSeriesComment:
    properties:
      createdAt:
        type: string
      depth:
        type: integer
      id:
        type: integer
      minusCount:
        type: integer
      parent:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesComment'
      parentId:
        type: integer
      plusCount:
        type: integer
      replies:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeriesComment'
        type: array
      repliesCount:
        type: integer
      text:
        type: string
      tvSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeries'
      tvSeriesId:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.User'
      userId:
        type: integer
    type: object
  github_com_movie-matcher_backend_internal_models.TVSeriesFavorite:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      tvSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeries'
      tvSeriesId:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.User'
      userId:
        type: integer
    type: object
  github_com_movie-matcher_backend_internal_models.TVSeriesList:
    properties:
      comment:
        type: string
      completedAt:
        type: string
      createdAt:
        type: string
      currentEpisode:
        type: integer
      currentProgress:
        type: number
      id:
        type: integer
      rating:
        type: integer
      startedAt:
        type: string
      status:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.ListStatus'
      titleReaction:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TitleReaction'
      tvSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeries'
      tvSeriesId:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.User'
      userId:
        type: integer
    type: object
  github_com_movie-matcher_backend_internal_models.TVSeriesReview:
    properties:
      actingRating:
        type: integer
      characterRating:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      musicRating:
        type: integer
      overallRating:
        type: integer
      productionRating:
        type: integer
      review:
        type: string
      reviewStatus:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.ReviewStatus'
      storyRating:
        type: integer
      tvSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeries'
      tvSeriesId:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.User'
      userId:
        type: integer
      visualsRating:
        type: integer
    type: object
  github_com_movie-matcher_backend_internal_models.TVSeriesSite:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      site:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.Site'
      siteId:
        type: integer
      tvSeries:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeries'
      tvSeriesId:
        type: integer
      updatedAt:
        type: string
      url:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_models.Theme:
    properties:
      createdAt:
        type: string
      description:
        type: string
      descriptionI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      emoji:
        type: string
      id:
        type: integer
      mediaTypes:
        items: {}
        type: array
      name:
        type: string
      nameI18n:
        $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.LocalizedString'
      updatedAt:
        type: string
    type: object
...
  github_com_movie-matcher_backend_internal_services.SemanticSearchItem:
    properties:
      description:
        type: string
      media_id:
        type: integer
      media_type:
        type: string
      poster:
        type: string
      score:
        type: number
      title:
        type: string
    type: object
  github_com_movie-matcher_backend_internal_services.SemanticSearchResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/github_com_movie-matcher_backend_internal_services.SemanticSearchItem'
        type: array
    type: object
...
  internal_handlers.ExtendedPaginatedResponse:
    properties:
      data: {}
      filters:
        $ref: '#/definitions/internal_handlers.FilterInfo'
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  internal_handlers.FilterInfo:
    properties:
      ageRatings:
        items:
          type: string
        type: array
      authorIds:
        items:
          type: string
        type: array
      countries:
        items:
          type: string
        type: array
      developerIds:
        items:
          type: string
        type: array
      genreIds:
        items:
          type: string
        type: array
      genreMode:
        type: string
      illustratorIds:
        items:
          type: string
        type: array
      maxRating:
        type: number
      minRating:
        type: number
      order:
        type: string
      platformIds:
        items:
          type: string
        type: array
      publisherIds:
        items:
          type: string
        type: array
      search:
        type: string
      sortBy:
        type: string
      studioIds:
        items:
          type: string
        type: array
      themeIds:
        items:
          type: string
        type: array
      themeMode:
        type: string
      year:
        type: integer
      yearFrom:
        type: integer
      yearTo:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /search/semantic:
    get:
      description: Search media by semantic similarity to the query (description,
        themes, mood). Requires recommendation service.
      parameters:
      - description: Search query (by meaning)
        in: query
        name: q
        required: true
        type: string
      - description: 'Filter: movie, anime, game, etc.'
        in: query
        name: mediaType
        type: string
      - description: Max results (default 20)
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_movie-matcher_backend_internal_services.SemanticSearchResponse'
      summary: Semantic search by description/meaning
      tags:
      - search
...
  /stream/video:
    get:
      parameters:
      - description: Path relative to UploadDir
        in: query
        name: path
        required: true
        type: string
      - description: 1080|720|480|360 — transcoded version
        in: query
        name: quality
        type: string
      responses:
        "200":
          description: video stream
        "400":
          description: error
        "404":
          description: error
      summary: Stream video file (trailers, etc.)
      tags:
      - Stream
...
  /tv-series/{id}:
    get:
      parameters:
      - description: TV Series ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeries'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get TV series by ID
      tags:
      - TV Series
  /tv-series/popular:
    get:
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: popular
        description: popular|trending|rating
        in: query
        name: mode
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_movie-matcher_backend_internal_models.TVSeries'
            type: array
      summary: Get popular TV series
      tags:
      - TV Series
  /tv-series/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ExtendedPaginatedResponse'
      summary: Search TV series
      tags:
      - TV Series
...
swagger: "2.0"